requirements/common.txt
Django==2.1
djangorestframework==3.8.2
django-extensions==2.1.0
psycopg2-binary
requests==2.19.1
xlwt==1.3.0
xlrd==1.1.0
gevent==1.3.5
redis==2.10.6
pytz==2018.5
ipython==6.5.0
django-model-utils==3.0.0
celery==4.1.0
msgpack-python==0.4.8
django-cors-headers==2.1.0
nickname_generator
django-anymail[mailgun]